+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 ## Changes
 - Changed application errors to make them more clear (#237)
 - Implemented the `x/upgrade` module (#174)
+- Renamed posts' parameters queries endpoints (#245)
 
 ## Bug fixes
 - Fixed a bug that caused valid URIs to be considered invalid (#233)
-
+ 
 # Version 0.9.0
 ## Changes
 - Removed the JSON-style `String` methods where it was possible, changed the others to not